fix: 修复语法错误

+++ lib/views/fourth_page/pager_indicator.dart
@@ -41,7 +41,7 @@ class PagerIndicator extends StatelessWidget {
             page.iconAssetPath,
             page.color,
             isHollow,
-            percentActive,
+            percentActive as double,
           ),
         ),
       );
